refactor: Deduplicate apply_deletions and promote KeyExistenceFilter to write module
