+++ rust/lance/src/dataset/write/merge_insert/exec.rs
@@ -4,19 +4,15 @@
 mod delete;
 mod write;
 
-use std::collections::BTreeMap;
-use std::sync::Arc;
-
 use datafusion::physical_plan::metrics::{Count, ExecutionPlanMetricsSet, MetricBuilder};
-use futures::StreamExt;
-use lance_table::format::Fragment;
-use roaring::RoaringTreemap;
 
 pub use delete::DeleteOnlyMergeInsertExec;
 pub use write::FullSchemaMergeInsertExec;
 
 use super::MergeStats;
-use crate::Dataset;
+
+// Re-export apply_deletions from the parent write module for use by exec submodules
+pub(super) use crate::dataset::write::apply_deletions;
 
 pub(super) struct MergeInsertMetrics {
     pub num_inserted_rows: Count,
@@ -59,50 +55,4 @@ impl MergeInsertMetrics {
             num_skipped_duplicates,
         }
     }
-}
-
-pub(super) async fn apply_deletions(
-    dataset: &Dataset,
-    removed_row_addrs: &RoaringTreemap,
-) -> crate::Result<(Vec<Fragment>, Vec<u64>)> {
-    let bitmaps = Arc::new(removed_row_addrs.bitmaps().collect::<BTreeMap<_, _>>());
-
-    enum FragmentChange {
-        Unchanged,
-        Modified(Box<Fragment>),
-        Removed(u64),
-    }
-
-    let mut updated_fragments = Vec::new();
-    let mut removed_fragments = Vec::new();
-
-    let mut stream = futures::stream::iter(dataset.get_fragments())
-        .map(move |fragment| {
-            let bitmaps_ref = bitmaps.clone();
-            async move {
-                let fragment_id = fragment.id();
-                if let Some(bitmap) = bitmaps_ref.get(&(fragment_id as u32)) {
-                    match fragment.extend_deletions(*bitmap).await {
-                        Ok(Some(new_fragment)) => {
-                            Ok(FragmentChange::Modified(Box::new(new_fragment.metadata)))
-                        }
-                        Ok(None) => Ok(FragmentChange::Removed(fragment_id as u64)),
-                        Err(e) => Err(e),
-                    }
-                } else {
-                    Ok(FragmentChange::Unchanged)
-                }
-            }
-        })
-        .buffer_unordered(dataset.object_store.io_parallelism());
-
-    while let Some(res) = stream.next().await.transpose()? {
-        match res {
-            FragmentChange::Unchanged => {}
-            FragmentChange::Modified(fragment) => updated_fragments.push(*fragment),
-            FragmentChange::Removed(fragment_id) => removed_fragments.push(fragment_id),
-        }
-    }
-
-    Ok((updated_fragments, removed_fragments))
 }
